// src/features/order/models/order.views/oneOrderDataView.model.ts
export class OneOrderDataViewModel {
  RAZN_KEY: number;
  REMONT: string;
  B_VOD: string;
  DATS: Date;
  KARTA: string;
  KARTAI: number;
  VR_V: Date;
  VR_Z: Date;
  RAZN_OD_ID: number;
  ZAPR: string;
  FIO_ID: number;
  NPL: string;
  COMMENTAR: string;
  RAZN_DL_ID: number;
  PRINT: string;
  DOP_DATE: number;
  END_DATE: Date;
  FIO_ID_KOND: number;
  ORG_ID: number;
  RAZN_OD_ID_PRICEP: number;
  FIO2_ID: number;
  TIP_PL: string;
  VID_PEREV: string;
  VID_SOOB: string;
}